// src/types/supabase.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      post: {
        Row: {
          content: string;
          cover: string;
          created_at: string;
          deleted_at: string | null;
          description: string;
          id: number;
          is_public: boolean;
          title: string;
          type: Database['public']['Enums']['post_type'];
          updated_at: string | null;
        };
        Insert: {
          content: string;
          cover?: string;
          created_at?: string;
          deleted_at?: string | null;
          description: string;
          id?: number;
          is_public?: boolean;
          title: string;
          type: Database['public']['Enums']['post_type'];
          updated_at?: string | null;
        };
        Update: {
          content?: string;
          cover?: string;
          created_at?: string;
          deleted_at?: string | null;
          description?: string;
          id?: number;
          is_public?: boolean;
          title?: string;
          type?: Database['public']['Enums']['post_type'];
          updated_at?: string | null;
        };
        Relationships: [];
      };
// ...
    };
// ...
  };
// ...
};

type PublicSchema = Database[Extract<keyof Database, 'public'>];

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema['Tables'] & PublicSchema['Views'])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions['schema']]['Tables'] &
        Database[PublicTableNameOrOptions['schema']]['Views'])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions['schema']]['Tables'] &
      Database[PublicTableNameOrOptions['schema']]['Views'])[TableName] extends {
      Row: infer R;
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema['Tables'] &
        PublicSchema['Views'])
    ? (PublicSchema['Tables'] &
        PublicSchema['Views'])[PublicTableNameOrOptions] extends {
        Row: infer R;
      }
      ? R
      : never
    : never;
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema['Enums']
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions['schema']]['Enums']
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions['schema']]['Enums'][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema['Enums']
    ? PublicSchema['Enums'][PublicEnumNameOrOptions]
    : never;